Ensure the worker processes do not crash when parsing invalid syntax
The original code **never** handled the cases where `@babel/core`'s
`transform` method throws an error, it just _happened_ to be that we
didn't hit this case (because babel tended not to `throw` during
parsing).

This updates to ensure that our worker will always resolve, sometimes
with `{ code, metadata }` and other times (when an error happens)
`{ error, stack }`. When we resolve with `{ error, stack }` the parent
process will throw an error with that information (so that `ember build`
exits non-zero, or `ember serve` shows the build failure and then
rebuilds succesfully once the syntax error is fixed).

(cherry picked from commit 91d78261206a606863b1198cc8965c3711487720)